# Programming_With_Typescript_And_Kotlin/code/jvm/http-api/src/main/kotlin/pt/isel/daw/model/output/OutputUser.kt
package pt.isel.daw.model.output

import pt.isel.daw.User

data class OutputUser (
    val id: Int,
    val name: String,
    val email: String,
)

fun User.toOutputUser() = OutputUser(
    id = id,
    name = name,
    email = email,
)